+++ src/app/components/pages/Home/Components/index.js
@@ -1,3 +1,4 @@
 export { default as SearchHeader } from './SearchHeader';
 export { default as SearchItem } from './SearchItem';
 export { default as SearchRelatedItem } from './SearchItem/SearchRelatedItem';
+export { default as Home } from './HomeContainer';
